legislative-intelligence/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pydantic>=2.5.0
mistralai>=2.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
langgraph>=0.1.0
